+++ nuxeo-drive-client/nxdrive/direct_edit.py
@@ -371,7 +371,7 @@ class DirectEdit(Worker):
     def _handle_lock_queue(self):
         local = self._local_client
 
-        while not self._lock_queue.empty():
+        while 'items':
             try:
                 item = self._lock_queue.get_nowait()
             except Empty:
@@ -389,6 +389,7 @@ class DirectEdit(Worker):
                     remote.lock(uid)
                     local.set_remote_id(dir_path, '1', name='nxdirecteditlock')
                     # Emit the lock signal only when the lock is really set
+                    self._send_lock_status(ref)
                     self.autolock.documentLocked.emit(os.path.basename(ref))
                     continue
 
@@ -399,7 +400,7 @@ class DirectEdit(Worker):
                 else:
                     purge = False
 
-                if purge or action.startswith('unlock'):
+                if purge or action == 'unlock_orphan':
                     path = local.abspath(ref)
                     log.trace('Remove orphan: %r', path)
                     self.autolock.orphan_unlocked(path)
@@ -408,13 +409,24 @@ class DirectEdit(Worker):
 
                 local.remove_remote_id(dir_path, name='nxdirecteditlock')
                 # Emit the signal only when the unlock is done
+                self._send_lock_status(ref)
                 self.autolock.documentUnlocked.emit(os.path.basename(ref))
             except ThreadInterrupt:
                 raise
             except:
                 # Try again in 30s
                 log.exception('Cannot %s document %r', action, ref)
-                self.directEditLockError.emit(action, os.path.basename(ref), uid)
+                self.directEditLockError.emit(
+                    action, os.path.basename(ref), uid)
+
+    def _send_lock_status(self, ref):
+        manager = self._manager
+        for engine in manager._engine_definitions:
+            dao = manager._engines[engine.uid]._dao
+            state = dao.get_states_from_remote(ref)
+            if state:
+                path = os.path.join(engine.local_folder, state.local_path)
+                manager.osi.send_sync_status(state, path)
 
     def _handle_upload_queue(self):
         local = self._local_client