+++ docs/changes.md
@@ -1,6 +1,11 @@
 # dev
 Release date: `2018-??-??`
 
+Changes in command line arguments:
+- `access-online`
+- `share-link` -> `copy-share-link`
+- `metadata` -> `edit-metadata`
+
 ### Core
 - [NXDRIVE-626](https://jira.nuxeo.com/browse/NXDRIVE-626): Use Crowdin for label translations
 - [NXDRIVE-925](https://jira.nuxeo.com/browse/NXDRIVE-925): Conflict resolve using local is not working
@@ -25,6 +30,8 @@ Release date: `2018-??-??`
 - [NXDRIVE-1136](https://jira.nuxeo.com/browse/NXDRIVE-1136): Change systray icon on update
 - [NXDRIVE-1140](https://jira.nuxeo.com/browse/NXDRIVE-1140): Use a GIF for the transferring icon
 - [NXDRIVE-1149](https://jira.nuxeo.com/browse/NXDRIVE-1149): New language: Indonesian
+- [NXDRIVE-1157](https://jira.nuxeo.com/browse/NXDRIVE-1157): Use file system decoration on macOS
+- [NXDRIVE-1158](https://jira.nuxeo.com/browse/NXDRIVE-1158): Restore the context menu "Edit metadata"
 - [NXDRIVE-1166](https://jira.nuxeo.com/browse/NXDRIVE-1166): Display a notification on new update on GNU/Linux
 - [NXDRIVE-1175](https://jira.nuxeo.com/browse/NXDRIVE-1175): New language: Hebrew
 
@@ -51,10 +58,12 @@ Release date: `2018-??-??`
 - Packaging: Added `requests` 2.18.4
 - Packaging: Removed `cffi`, will be installed with `xattr`
 - Packaging: Updated `Js2Py` from 0.58 to 0.59
+- Packaging: Updated `pycryptodomex` from 3.5.1 to 3.6.0
 - Packaging: Updated `pypac` from 0.4.0 to 0.8.1
 - Packaging: Updated `pytest` from 3.3.2 to 3.5.0
 - Packaging: Updated `python-dateutil` from 2.6.1 to 2.7.2
 - Packaging: Updated `xattr` from 0.9.2 to 0.9.3
+- Packaging: Upgraded `SIP` from 4.19.7 to 4.19.8
 - Tracker: Removed the `Update` event